requirements.txt
earthengine-api
google-auth
dotenv
rasterio
numpy
opencv-python
shapely
geopandas
pyproj
google-cloud-storage
tensorflow
fastapi
uvicorn